renamed auth.py to _internal_auth.py to indicate it is internal
