+++ python/delta_sharing/tests/test_oauth_client.py
@@ -18,7 +18,7 @@ import requests
 from requests.models import Response
 from unittest.mock import patch
 from datetime import datetime
-from delta_sharing.auth import OAuthClient
+from delta_sharing._internal_auth import OAuthClient
 
 
 class MockServer:
